SALTO-7520: Ignore merge errors between nacl and state (#17)
When we encounter an issue that causes hidden values to be written to the nacl, we get merge errors
Currently merge errors cause the workspace to be considered errored and block many things.
We assume that this behavior is much more likely to happen due to a bug than a user action, so we ignore these merge errors and only log about them.
Using the value from the state to maintain the invariant that hidden values cannot be changed directly (we never deploy a hidden value)


In this change we reduce the severity of merge errors from that step to warning, so they would still be there, but they won't be blocking