SALTO-7335: Ignoring keys from the Object prototype in multi index (#198)
If `get` from a multi index is called with a key from the Object prototype (such as "constructor") and it isn't present in the index, it will return a function instead of the value we want. This can cause all sorts of bugs (and has).
Introducing the new `getOwn` helper to lowerdash to solve this. We should probably use this utility instead of `_.get` everywhere, but that is way out of the scope of this PR.